part-7/dev-requirements.txt
asgiref==3.5.0
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
click==8.0.4
Django==4.0.3
django-elasticsearch-dsl==7.2.2
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
